moved test utils code to conftest
