+++ imap_processing/tests/idex/conftest.py
@@ -1,10 +1,105 @@
 from pathlib import Path
+from typing import Optional
 
+import h5py
+import numpy as np
 import pytest
 import xarray as xr
 
 from imap_processing import imap_module_directory
 from imap_processing.idex.idex_l1a import PacketParser
+
+
+def load_hdf_file(path: str, num_events: Optional[int] = None) -> dict:
+    """
+    Loads an HDF5 file produced by the IDEX team into a dictionary where the keys
+    correspond to the groups in the file, and the values are arrays containing the data.
+
+    The HDF5 files are organized by event number, where each event contains groups
+    for metadata and data items. This function consolidates all events into a single
+    array to be consistent with the cdf organization.
+    For example, if there are 14 events and the metadata value "category" is extracted,
+    it will appear in the output dictionary as:
+    {"category": np.ndarray of shape (14,)}.
+
+    Parameters:
+    ----------
+    path : str
+        The file path to the HDF5 file.
+    num_events : int, optional
+        The number of events to extract. If None, all events in the file are processed.
+
+    Returns:
+    -------
+    dict
+        A dictionary containing the extracted data.
+
+    Warnings:
+    --------
+    - This function assumes the HDF5 file is structured with top-level groups named
+      numerically (e.g., "1", "2", "3") representing events.
+    """
+
+    def collect_arrays(name, obs):
+        """
+        Collects data from HDF5 datasets into the dictionary `data_vars`.
+
+        Parameters:
+        ----------
+        name : str
+            The full path of the dataset in the HDF5 file.
+        obs : h5py.Dataset or h5py.Group
+            The object at the current position in the HDF5 file.
+        """
+        # Split the full dataset path to extract the group key and event number
+        names = name.split("/")
+        group_key = names[-1]
+        event = names[0]
+
+        # Check if the event name can be converted to an integer
+        # This function assumes all top level groups are event numbers represented as
+        # strings
+        try:
+            event_number = int(event)
+        except ValueError as e:
+            raise ValueError(
+                f"Invalid group name '{event}': Top-level groups must be integers."
+            ) from e
+
+        # Process data only if the event number is within the specified range
+        if event_number <= num_events:
+            if isinstance(obs, h5py.Dataset):
+                # Initialize storage for the group key if not already present
+                if group_key not in data_vars.keys():
+                    # handle arrays
+                    if isinstance(obs[()], np.ndarray):
+                        data_vars[group_key] = np.zeros(
+                            (num_events, len(obs[()])), dtype=type(obs[()])
+                        )
+                    # Handle byte strings
+                    elif isinstance(obs[()], bytes):
+                        data_vars[group_key] = np.zeros(num_events, dtype=object)
+                    # Handle other types (scalars)
+                    else:
+                        data_vars[group_key] = np.zeros(num_events, dtype=obs[()].dtype)
+                # Convert bytes to string
+                if isinstance(obs[()], bytes):
+                    value = obs.asstr()[()]
+                else:
+                    value = obs[()]
+                # store value at event number
+                data_vars[group_key][int(event) - 1] = value
+
+    # Open the HDF5 file
+    f = h5py.File(path, "r")
+    if not num_events or num_events > len(f.keys()):
+        num_events = len(f.keys())
+    # Initialize dictionary
+    data_vars = {}
+    # First extract metadata from each event
+    f.visititems(collect_arrays)
+
+    return data_vars
 
 
 @pytest.fixture()
@@ -20,3 +115,22 @@ def decom_test_data() -> xr.Dataset:
         f"{imap_module_directory}/tests/idex/test_data/imap_idex_l0_raw_20231214_v001.pkts"
     )
     return PacketParser(test_file, "001").data
+
+
+@pytest.fixture()
+def l1a_example_data():
+    """
+    Pytest fixture to load example L1A data (produced by the IDEX team) for testing.
+
+    Returns
+    -------
+    dict
+      A dictionary containing the 6 waveform and telemetry arrays
+    """
+    l1a_test_path = (
+        f"{imap_module_directory}/tests/idex/validation_files/L1A_Example.h5"
+    )
+    # The H5 files were modified to include only the first 10 events to ensure the file
+    # size stays below the maximum limit of 1MB.
+    num_events = 7
+    return load_hdf_file(l1a_test_path, num_events=num_events), num_events
