+++ imap_processing/tests/idex/test_idex_l1a.py
@@ -7,26 +7,6 @@ from cdflib.xarray.xarray_to_cdf import ISTPError
 
 from imap_processing import imap_module_directory
 from imap_processing.cdf.utils import load_cdf, write_cdf
-from imap_processing.tests.idex.test_utils import load_hdf_file
-
-
-@pytest.fixture()
-def l1a_example_data():
-    """
-    Pytest fixture to load example L1A data (produced by the IDEX team) for testing.
-
-    Returns
-    -------
-    dict
-      A dictionary containing the 6 waveform and telemetry arrays
-    """
-    l1a_test_path = (
-        f"{imap_module_directory}/tests/idex/validation_files/L1A_Example.h5"
-    )
-    # The H5 files were modified to include only the first 10 events to ensure the file
-    # size stays below the maximum limit of 1MB.
-    num_events = 7
-    return load_hdf_file(l1a_test_path, num_events=num_events), num_events
 
 
 def test_idex_cdf_file(decom_test_data: xr.Dataset):
@@ -108,7 +88,7 @@ def test_idex_tof_high_data_from_cdf(decom_test_data: xr.Dataset):
     ) as f:
         data = np.array([int(line.rstrip()) for line in f])
 
-    file_name = write_cdf(decom_test_data, istp=False)
+    file_name = write_cdf(decom_test_data)
     l1_data = load_cdf(file_name)
     assert (l1_data["TOF_High"][13].data == data).all()
 
@@ -120,6 +100,10 @@ def test_validate_l1a_idex_data_variables(
     Verify that each of the 6 waveform and telemetry arrays are equal to the
     corresponding array produced by the IDEX team using the same l0 file.
 
+    The comparison is limited to `num_events` because the L1A example contains fewer
+    events (due to file size requirements) than the SCD dataset.
+
+
     Parameters
     ----------
     decom_test_data : xarray.Dataset
@@ -128,26 +112,37 @@ def test_validate_l1a_idex_data_variables(
         A dictionary containing the 6 waveform and telemetry arrays
     """
     # Lookup table to match the SDS array names to the Idex Team array names
-    waveform_variables = {
-        "TOF_Low": "TOF L",
-        "TOF_High": "TOF H",
-        "TOF_Mid": "TOF M",
-        "Target_High": "Target H",
-        "Target_Low": "Target L",
-        "Ion_Grid": "Ion Grid",
+    l1a_examples = l1a_example_data[0]
+    # Number of events in the l1a_examples dict for each data_variable
+    num_events = l1a_example_data[1]
+    match_variables = {
+        "TOF L": "TOF_Low",
+        "TOF H": "TOF_High",
+        "TOF M": "TOF_Mid",
+        "Target H": "Target_High",
+        "Target L": "Target_Low",
+        "Ion Grid": "Ion_Grid",
+        "Time (high sampling)": "time_high_sr",
+        "Time (low sampling)": "time_low_sr",
     }
 
-    for var in decom_test_data.data_vars:
-        l1a_examples = l1a_example_data[0]
-        num_events = l1a_example_data[1]
-        if var in waveform_variables.keys():
-            eng_var = waveform_variables[var]
-        else:
-            eng_var = var.upper()
+    # The Engineering data is converting to UTC, and the SDC is converting to J2000,
+    # for 'epoch' and 'Timestamp' so this test is using the raw time value 'SCHOARSE' to
+    # validate time
+    arrays_to_skip = ["Timestamp", "Epoch"]
 
-        assert np.array_equal(
-            decom_test_data[var][:num_events], l1a_examples[eng_var]
-        ), (
-            f"The array '{var}' does not equal the expected example array '{eng_var}' "
-            f"produced by the IDEX team"
-        )
+    # loop through all keys from the l1a example dict
+    for var in l1a_examples.keys():
+        if var not in arrays_to_skip:
+            # Find the corresponding array name
+            if var in match_variables.keys():
+                cdf_var = match_variables[var]
+            else:
+                cdf_var = var.lower()
+
+            assert np.array_equal(
+                decom_test_data[cdf_var][:num_events], l1a_examples[var]
+            ), (
+                f"The array '{cdf_var}' does not equal the expected example array "
+                f"'{var}' produced by the IDEX team"
+            )
